backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]   # para manejar JWT
passlib[bcrypt]             # para hashear/verificar contraseñas